Optimize ActiveNode deserialization (#1490)
* Optimize TypeConverters#skip_conversion?

`obj.class.attributes.[].!@` is four method calls and is unlikely to
short-circuit the method unless the DB node has properties that the Ruby
class doesn't declare. This commit prioritizes the condition that
includes a single method call and is more likely to short-circuit the
conditional.

* Optimize DeclaredProperty assignment

Assigning to instance variables instead of pulling from a hash at
time of assignment increases performance since it doesn't have to
calculate a hash bucket each time.

* Remove unnecessary scoping

Constant lookup was being done 2x for each call to `converted_property`.

* Use a set for inclusion check

Sets are faster than an array when there are more than 3-5 elements

* Swap order of checks to improve short-circuit rate

* Check hash inclusion rather than set inclusion

This happens in O(1) time rather than O(n) and removes the intermediate
array.

* Let things be options

Hopefully this fixes CI

* Fix type/options pluck that I left in

* Options are in the public API, so update them

* Turn this back into a hash

* Remove options[:type] and friends

We should be relying on the method rather than values in the options
hash.